Support to disable keyseparator.
Add simple parser tests to check keyseparator behaviour.

DRY test moving custom transform to util module.

Trim whitespaces.
