+++ lib/parser.js
@@ -121,7 +121,7 @@ var Parser = function(options) {
  * @param {string} key The translation key
  * @param {string} [defaultValue] The key's value
  */
-Parser.prototype.parse = function(key, defaultValue) { 
+Parser.prototype.parse = function(key, defaultValue) {
     var fn = this.fn;
     var options = this.options;
     var resStore = this.resStore;
@@ -135,7 +135,11 @@ Parser.prototype.parse = function(key, defaultValue) {
         key = parts[1];
     }
 
-    var keys = key.split(options.keyseparator);
+    // keep compatibility with i18next to supoprt disabling keyseparator
+    // in case you want to have the key as a fallback value.
+    var keys = typeof options.keyseparator === 'string' ?
+                key.split(options.keyseparator) :
+                [key];
     _.each(options.lngs, function(lng) {
         var lookupKey;
         var value = resStore[lng] && resStore[lng][ns];
@@ -148,16 +152,16 @@ Parser.prototype.parse = function(key, defaultValue) {
         if ( ! _.isUndefined(value)) {
             // Found a value associated with the key
             lookupKey = '[' + lng + '][' + ns + '][' + keys.join('][') + ']';
-            fn.debuglog('Found a value %s associated with the key %s in %s.', 
+            fn.debuglog('Found a value %s associated with the key %s in %s.',
                 JSON.stringify(_.get(resStore, lookupKey)),
-                JSON.stringify(keys.join(options.keyseparator)),
+                JSON.stringify(key),
                 JSON.stringify(fn.getResourcePath(lng, ns))
             );
         } else if (_.isObject(resStore[lng][ns])) {
             // Adding a new entry
             var res = resStore[lng][ns];
             _.each(keys, function(elem, index) {
-                if (index >= (keys.length - 1)) { 
+                if (index >= (keys.length - 1)) {
                     res[elem] = _.isUndefined(defaultValue) ? options.defaultValue : defaultValue;
                 } else {
                     res[elem] = res[elem] || {};
@@ -167,7 +171,7 @@ Parser.prototype.parse = function(key, defaultValue) {
 
             lookupKey = '[' + lng + '][' + ns + '][' + keys.join('][') + ']';
             fn.debuglog('Adding a new entry {%s:%s} to %s.',
-                JSON.stringify(keys.join(options.keyseparator)),
+                JSON.stringify(key),
                 JSON.stringify(_.get(resStore, lookupKey)),
                 JSON.stringify(fn.getResourcePath(lng, ns))
             );
@@ -226,7 +230,7 @@ Parser.prototype.parseAttrs = function(attrs) {
  * @example <caption>Example usage:</caption>
  * // it will output ["id=nav-bar", "class = "top"", "foo = "bar\"baz""]
  * var str = ' id=nav-bar class = "top" foo = "bar\\"baz" ';
- * str.match(/([^=,\s]*)\s*=\s*((?:"(?:\\.|[^"\\]+)*"|'(?:\\.|[^'\\]+)*')|[^'"\s]*)/igm) || [];              
+ * str.match(/([^=,\s]*)\s*=\s*((?:"(?:\\.|[^"\\]+)*"|'(?:\\.|[^'\\]+)*')|[^'"\s]*)/igm) || [];
  * @param [string] str A string representation of hash arguments
  * @return {object}
  */
@@ -288,7 +292,7 @@ Parser.prototype.toJSON = function(opts) {
  * Returns an Object containing translation information
  * @param {object} [opts] The options object
  * @param {boolean} [opts.sort] True to sort object by key
- * @return {object} 
+ * @return {object}
  */
 Parser.prototype.toObject = function(opts) {
     opts = opts || {};
